Actualizacion del README, creacion de TDA server y client. Chequeo de endianess

+++ common_socket.h
@@ -2,14 +2,7 @@
 #define COMMON_SOCKET_H
 
 #include <sys/types.h>
-#include <sys/socket.h>
-#include <sys/un.h>
 #include <netdb.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <string.h>
-#include <stdbool.h>
 
 #include "common_mod26.h"
 
@@ -23,9 +16,9 @@ void socket_init(socket_t *this);
 
 void socket_connect(socket_t *this, const char *service, const char *port);
 
-void socket_send(socket_t *this, const char *data, int l);
+void socket_send(socket_t *this, const char *data, uint16_t l);
 
-int socket_read(socket_t *this, char *data);
+uint16_t socket_read(socket_t *this, char *data);
 
 void socket_uninit(socket_t *this);
